requirements.txt
streamlit==1.39.0
pandas>=2.2.0
numpy>=1.26.0
torch==2.7.1
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
Pillow>=10.2.0
psutil>=5.9.0
